Changed paths: M       CHANGES.txt M       cat_main.m M       cat_main1639.m M       cat_main_LASs.m M       cat_main_amap.m M       cat_main_gintnorm.m M       cat_main_gintnormi.m M       cat_main_registration.m M       cat_main_updateSPM.m M       cat_main_updateSPM1639.m M       cat_run.m M       cat_run_job.m M       cat_run_job1639.m M       cat_surf_fun.m M       cat_surf_render2.m M       cat_vol_iscaling.m M       cat_vol_partvol.m M       cat_vol_qa.m M       cat_vol_sanlm.m
(1) Fixed:   Tested the new cat_main pipeline in case of non-T1 contrasts.
             Corrected bugs in the call and with the boundary box in cat_main. 
             Added simple skull-stripping for non-T1 data in cat_main_updateSPM.
             Corrected small problems PD/T2 processing in cat_vol_partvol.
             Updated cat_main_gintnorm with k-means peaks rather than SPM peaks
             in the backup/T2/PD stream - it is ok but could be better.  
             Added high accuracy (smaller spm.tol values) in cases of failed 
             SPM preprocessing 1 in cat_run_job.
             Updated simple LAS function but it is still unstable and turned off.  
(2) Fixed:   Corrected bug in cat_vol_sanlm for images with negative values.               
(3) Fixed:   Corrected bug of matrix size for different template size in 
             cat_main_registration.  
(4) Fixed:   Corrected resolution rounding issue in cat_run[1639].
(5) New:     Added output of unprocessed files.         
(6) Fixed:   Tiny corrections in cat_vol_qa,       
(7) Fixed:   Corrected typo in cat_surf_render2.
              


git-svn-id: svn://141.35.69.214/cat12@1834 9f44b166-36c4-483c-8295-4af200601734

+++ cat_main_updateSPM1639.m
@@ -91,7 +91,7 @@ function [Ysrc,Ycls,Yb,Yb0,job,res,T3th,stime2] = cat_main_updateSPM1639(Ysrc,P,
   else
     CMth = min( [  clsint(1) - diff([clsint(1),WMth]) , clsint(3) ]);
   end
-  T3th = [ CMth , clsint(1) , WMth];
+  T3th = double([ CMth , clsint(1) , WMth]);
 
 
   %% Some error handling
